requirements.txt
geopy==2.4.1
mysql-connector-python==8.0.33
requests==2.32.5
Unidecode==1.4.0
flask-cors==6.0.1
